requirements.txt
aiosmtplib==5.0.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
blinker==1.9.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.127.0
fastapi-mail==1.6.1
greenlet==3.3.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
phonenumbers==9.0.21
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
regex==2025.11.3
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
